clean up utility files added
