+++ packages/react-storage/src/components/StorageBrowser/Views/Controls/constants.ts
@@ -1,2 +1,9 @@
 export const ACTION_MENU_BLOCK_NAME = 'action-menu';
 export const ACTION_MENU_NAME = 'Action menu';
+export const ICON_ATTRIBUTES = {
+  'aria-hidden': true,
+  width: '24',
+  height: '24',
+  viewBox: '0 0 24 24',
+  xmlns: 'http://www.w3.org/2000/svg',
+};
